services/rest/requirements.txt
django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf_yasg==1.21.3